2896: Convert ApplicationHandler to ApplicationService Spring bean
